Fix image upload to not delete default image
Also removed redundant logo upload and instead point to static files for
default image.

Closes: #57

+++ clicker_app/views.py
@@ -215,7 +215,9 @@ def upload_image(request):
                 new_image = image_form.files["image"]
                 user_account = Account.objects.get(user__id=account_id)
                 old_image = user_account.image
-                old_image.delete()
+                if str(old_image) != "../static/images/logo.png":
+                    old_image.delete()
+                # TODO: maybe change image name to account_id new_image.name = account_id
                 user_account.image = new_image
                 user_account.save()
